requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.0
pydantic==2.5.2
python-multipart==0.0.6
httpx==0.24.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.4.0
passlib==1.7.4
bcrypt==4.0.1 